Initial commit of Platform Engineer interview task - lite
Why:
Marysonia created an interview task for platform engineers we were looking to hire.
The full task takes several hours and is too long for hiring contractors that already
go through a full interview process with the contracting company.
This is meant to be a shorter version of the platform engineering interview questions.

What:
Removed several exercises from the full version to make this shorter